feat: improve DynamicConfigForm * Display current config value from store * Add `path` form item type
