+++ src/components/Nodes/ClientConfig/index.js
@@ -7,7 +7,6 @@ import Tabs from '@material-ui/core/Tabs'
 import Tab from '@material-ui/core/Tab'
 import Typography from '@material-ui/core/Typography'
 import VersionList from './VersionList'
-// import ConfigForm from './ConfigForm'
 import DynamicConfigForm from './DynamicConfigForm'
 import Terminal from '../Terminal'
 // import NodeInfo from '../NodeInfo'
@@ -144,6 +143,7 @@ class ClientConfig extends Component {
         {activeTab === 1 && (
           <TabContainer>
             <DynamicConfigForm
+              clientName={client.name}
               settings={settings}
               handleClientConfigChanged={this.handleClientConfigChanged}
               isClientRunning={isRunning}
